refactor: 메시지 얻는 로직 함수로 분리
