+++ src/components/common/ErrorMessage/ErrorMessage.tsx
@@ -2,5 +2,6 @@ import * as S from './ErrorMessage.styles';
 import { PropsWithChildren } from 'react';
 
 export default function ErrorMessage({ children }: PropsWithChildren) {
+  if (!children) return null;
   return <S.ErrorBox>{children}</S.ErrorBox>;
 }
